move transport2030 g to energy_general
